requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
httpx>=0.24.0
jsonschema>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
